+++ src/provider/Def/getDefWithLabel.ts
@@ -9,7 +9,7 @@ import { getHotkeyWrap } from '../../tools/Command/HotkeyTools';
 import { getMenuFunc } from '../../tools/Command/MenuTools';
 import { getSetTimerWrap } from '../../tools/Command/SetTimerTools';
 import type { TScanData } from '../../tools/DeepAnalysis/FnVar/def/spiltCommandAll';
-import { findLabel } from '../../tools/labels';
+import { findLabelAll } from '../../tools/labelsAll';
 import { ToUpCase } from '../../tools/str/ToUpCase';
 
 function LabelRefHotkey(AhkTokenLine: TAhkTokenLine, wordUp: string): vscode.Range | null {
@@ -112,44 +112,64 @@ function LabelRefKeyWordWrap(
     return [];
 }
 
-function getLabelRef(wordUp: string): vscode.Location[] {
-    // TODO for performance use keyword match replace regex!
-    const List: vscode.Location[] = [];
-    for (const { DocStrMap, uri } of pm.getDocMapValue()) {
-        for (const AhkTokenLine of DocStrMap) {
-            const { detail, lStr, line } = AhkTokenLine;
-            if (detail.includes(EDetail.isLabelLine) && ToUpCase(lStr.trim()) === (`${wordUp}:`)) {
-                List.push(
-                    new vscode.Location(
-                        uri,
-                        new vscode.Range(
-                            new vscode.Position(line, lStr.search(/\S/u)),
-                            new vscode.Position(line, lStr.indexOf(':')),
-                        ),
+type TFileRefMap = Map<string, readonly vscode.Location[]>;
+const fileLabelRefMap = new WeakMap<TAhkFileData, TFileRefMap>();
+
+// TODO remove wordUp param
+function getLabelRefCore(AhkFileData: TAhkFileData, wordUp: string): readonly vscode.Location[] {
+    const fileList: vscode.Location[] = [];
+    const { DocStrMap, uri } = AhkFileData;
+    for (const AhkTokenLine of DocStrMap) {
+        const { detail, lStr, line } = AhkTokenLine;
+        if (detail.includes(EDetail.isLabelLine) && ToUpCase(lStr.trim()) === (`${wordUp}:`)) {
+            fileList.push(
+                new vscode.Location(
+                    uri,
+                    new vscode.Range(
+                        new vscode.Position(line, lStr.search(/\S/u)),
+                        new vscode.Position(line, lStr.indexOf(':')),
                     ),
-                );
-                continue;
-            }
-            const range: vscode.Range | null = LabelRefHotkey(AhkTokenLine, wordUp);
-            if (range !== null) {
-                List.push(new vscode.Location(uri, range));
-                continue;
-            }
-
-            const menuRange: vscode.Range | null = LabelRefMenu(AhkTokenLine, wordUp);
-            if (menuRange !== null) {
-                List.push(new vscode.Location(uri, menuRange));
-                continue;
-            }
-
-            const range2: vscode.Range | null = LabelRefGui(AhkTokenLine, wordUp);
-            if (range2 !== null) {
-                List.push(new vscode.Location(uri, range2));
-                continue;
-            }
-
-            List.push(...LabelRefKeyWordWrap(wordUp, AhkTokenLine, uri));
+                ),
+            );
+            continue;
         }
+        const range: vscode.Range | null = LabelRefHotkey(AhkTokenLine, wordUp);
+        if (range !== null) {
+            fileList.push(new vscode.Location(uri, range));
+            continue;
+        }
+
+        const menuRange: vscode.Range | null = LabelRefMenu(AhkTokenLine, wordUp);
+        if (menuRange !== null) {
+            fileList.push(new vscode.Location(uri, menuRange));
+            continue;
+        }
+
+        const range2: vscode.Range | null = LabelRefGui(AhkTokenLine, wordUp);
+        if (range2 !== null) {
+            fileList.push(new vscode.Location(uri, range2));
+            continue;
+        }
+
+        fileList.push(...LabelRefKeyWordWrap(wordUp, AhkTokenLine, uri));
+    }
+
+    return fileList;
+}
+
+function getLabelRef(wordUp: string): vscode.Location[] {
+    const List: vscode.Location[] = [];
+    for (const AhkFileData of pm.getDocMapValue()) {
+        const map: TFileRefMap = fileLabelRefMap.get(AhkFileData) ?? new Map<string, readonly vscode.Location[]>();
+        const cache: readonly vscode.Location[] | undefined = map.get(wordUp);
+        if (cache !== undefined) {
+            List.push(...cache);
+            continue;
+        }
+        const fileList: readonly vscode.Location[] = getLabelRefCore(AhkFileData, wordUp);
+        map.set(wordUp, fileList);
+        fileLabelRefMap.set(AhkFileData, map);
+        List.push(...fileList);
     }
 
     return List;
@@ -160,22 +180,12 @@ export function posAtLabelDef(
     position: vscode.Position,
     wordUp: string,
 ): vscode.Location[] | null {
-    const { lStr } = AhkFileData.DocStrMap[position.line];
+    const { lStr, detail } = AhkFileData.DocStrMap[position.line];
 
-    if ((/^[#$@\w\u{A1}-\u{FFFF}]+:$/u).test(lStr.trim())) {
+    if (detail.includes(EDetail.isLabelLine) && (/^[#$@\w\u{A1}-\u{FFFF}]+:$/u).test(lStr.trim())) {
         return getLabelRef(wordUp);
     }
     return null;
-}
-
-function getDefWithLabelCore(wordUpCase: string): vscode.Location[] | null {
-    const label: CAhkLabel | null = findLabel(wordUpCase);
-    if (label === null) {
-        return null;
-    }
-    const { range, uri } = label;
-    const Location: vscode.Location = new vscode.Location(uri, range);
-    return [Location];
 }
 
 /**
@@ -189,19 +199,18 @@ function getDefWithLabelCore(wordUpCase: string): vscode.Location[] | null {
  */
 export function getDefWithLabel(
     AhkFileData: TAhkFileData,
-    uri: vscode.Uri,
     position: vscode.Position,
     wordUpCase: string,
-): vscode.Location[] | null {
+): CAhkLabel[] | null {
     const { DocStrMap } = AhkFileData;
 
     const { line, character } = position;
     const AhkTokenLine: TAhkTokenLine = DocStrMap[line];
-    const { lStr } = AhkTokenLine;
+    const { lStr, detail } = AhkTokenLine;
     const lStrFix: string = lStr.slice(0, Math.max(0, character));
 
-    if ((/^[#$@\w\u{A1}-\u{FFFF}]+:$/u).test(lStr.trim())) {
-        return [new vscode.Location(uri, position)]; // let auto call Ref
+    if (detail.includes(EDetail.isLabelLine) && (/^[#$@\w\u{A1}-\u{FFFF}]+:$/u).test(lStr.trim())) {
+        return findLabelAll(wordUpCase);
     }
 
     const {
@@ -215,26 +224,26 @@ export function getDefWithLabel(
         list.includes(fistWordUp)
         && character > fistWordUp.length + fistWordUpCol
         && (/[#$@\w\u{A1}-\u{FFFF}]*$/iu).test(lStrFix)
-    ) return getDefWithLabelCore(wordUpCase);
+    ) return findLabelAll(wordUpCase);
 
     if (
         list.includes(SecondWordUp)
         && character > SecondWordUp.length + SecondWordUpCol
         && (/[#$@\w\u{A1}-\u{FFFF}]*$/iu).test(lStrFix)
-    ) return getDefWithLabelCore(wordUpCase);
+    ) return findLabelAll(wordUpCase);
 
     const Data: TScanData | null = getHotkeyWrap(AhkTokenLine)
         ?? getMenuFunc(AhkTokenLine)
         ?? getSetTimerWrap(AhkTokenLine); // don't find `Sort`
 
-    if (ToUpCase(Data?.RawNameNew ?? '') === wordUpCase) return getDefWithLabelCore(wordUpCase);
+    if (ToUpCase(Data?.RawNameNew ?? '') === wordUpCase) return findLabelAll(wordUpCase);
 
     const GuiDataList: readonly TScanData[] | null = getGuiFunc(AhkTokenLine, 0);
     if (GuiDataList !== null) {
         const wordUpCaseFix: string = wordUpCase.replace(/^g/iu, '');
         for (const { RawNameNew } of GuiDataList) {
             if (ToUpCase(RawNameNew) === wordUpCaseFix) {
-                return getDefWithLabelCore(wordUpCaseFix);
+                return findLabelAll(wordUpCaseFix);
             }
         }
     }
@@ -242,6 +251,20 @@ export function getDefWithLabel(
     return null;
 }
 
+export function getDefWithLabelWrap(
+    AhkFileData: TAhkFileData,
+    uri: vscode.Uri,
+    position: vscode.Position,
+    wordUpCase: string,
+): vscode.Location[] | null {
+    const list: CAhkLabel[] | null = getDefWithLabel(
+        AhkFileData,
+        position,
+        wordUpCase,
+    );
+    if (list === null) return null;
+    return list.map((l: CAhkLabel): vscode.Location => new vscode.Location(l.uri, l.selectionRange));
+}
 // unknown...
 // {_T("GroupAdd"), 1, 6, 6, NULL} // Group name, WinTitle, WinText, Label, exclude-title/text
 // {_T("Menu"), 2, 6, 6, NULL}  // tray, add, name, label, options, future use